MAA-44 : adb key event doesn't want to work without changing path
Eclipse IDE was not able to pick up the ADB path from the MAC OS , hence it was explicitly set in the configuration properties to execute the ADB commands and key events. This issue is resolved since we have migrated to Intellij IDE.

The adb path has been removed from the configuration property.
Added the comments for the methods